requirements.txt
argcomplete==1.12.3
argon2-cffi==21.1.0
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
bleach==4.1.0
cffi==1.15.0
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==2.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
entrypoints==0.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.3
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.5
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
packaging==21.2
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
prometheus-client==0.12.0
prompt-toolkit==3.0.22
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pywinpty==1.1.5
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.6.0
